Support for KLC in localappdata, bookmarks, Canary+Charm. Added Kaseya SaaS KLC version check.

+++ Utilities/ConfigureHandler.cs
@@ -46,25 +46,45 @@ namespace KLCProxy
 
             InitializeCurrentUserValue();
 
-            RegistryKey subkey = Registry.ClassesRoot.OpenSubKey(@"liveconnect\shell\open\command", false);
-            RegistryKey subkeyCU = Registry.CurrentUser.OpenSubKey(@"Software\Classes\liveconnect\shell\open\command", true);
+            RegistryKey subkey1 = Registry.ClassesRoot.OpenSubKey(@"liveconnect\shell\open\command", false);
+            RegistryKey subkey1CU = Registry.CurrentUser.OpenSubKey(@"Software\Classes\liveconnect\shell\open\command", true);
 
-            if (enabled) {
+            RegistryKey subkey2 = Registry.ClassesRoot.OpenSubKey(@"kaseyaliveconnect\shell\open\command", false);
+            RegistryKey subkey2CU = Registry.CurrentUser.OpenSubKey(@"Software\Classes\kaseyaliveconnect\shell\open\command", true);
+
+            if (enabled)
+            {
                 if (bypass)
                 {
-                    subkeyCU.SetValue("", ExpectedValueFinch());
+                    subkey1CU.SetValue("", ExpectedValueFinch());
+                    subkey2CU.SetValue("", ExpectedValueFinch());
                     value = ProxyState.BypassToFinch;
                 }
                 else
                 {
-                    subkeyCU.SetValue("", ExpectedValue());
+                    subkey1CU.SetValue("", ExpectedValue());
+                    subkey2CU.SetValue("", ExpectedValue());
                     value = ProxyState.Enabled;
                 }
-            } else
-                Registry.CurrentUser.DeleteSubKeyTree(@"Software\Classes\liveconnect\shell\open");
+            }
+            else
+            {
+                string pathLocal = Environment.ExpandEnvironmentVariables(@"%localappdata%\Apps\Kaseya Live Connect\KaseyaLiveConnect.exe");
+                if (File.Exists(pathLocal))
+                {
+                    subkey1CU.SetValue("", string.Format("\"{0}\" \"%1\"", pathLocal));
+                    subkey2CU.SetValue("", string.Format("\"{0}\" \"%1\"", pathLocal));
+                } else
+                {
+                    Registry.CurrentUser.DeleteSubKeyTree(@"Software\Classes\liveconnect\shell\open");
+                    Registry.CurrentUser.DeleteSubKeyTree(@"Software\Classes\kaseyaliveconnect\shell\open");
+                }
+            }
 
-            subkey.Close();
-            subkeyCU.Close();
+            subkey1.Close();
+            subkey1CU.Close();
+            subkey2.Close();
+            subkey2CU.Close();
 
             return value;
         }
@@ -74,6 +94,20 @@ namespace KLCProxy
             if (Registry.CurrentUser.OpenSubKey(@"Software\Classes\liveconnect\shell\open\command", false) == null)
             {
                 using (RegistryKey key = Registry.CurrentUser.CreateSubKey("SOFTWARE\\Classes\\liveconnect"))
+                {
+                    key.SetValue("", "URL:Live Connect Protocol");
+                    key.SetValue("URL Protocol", "");
+
+                    using (var commandKey = key.CreateSubKey(@"shell\open\command"))
+                    {
+                        commandKey.SetValue("", "");
+                    }
+                }
+            }
+
+            if (Registry.CurrentUser.OpenSubKey(@"Software\Classes\kaseyaliveconnect\shell\open\command", false) == null)
+            {
+                using (RegistryKey key = Registry.CurrentUser.CreateSubKey("SOFTWARE\\Classes\\kaseyaliveconnect"))
                 {
                     key.SetValue("", "URL:Live Connect Protocol");
                     key.SetValue("URL Protocol", "");
